refactor: 사용안하는 파일 제거 및 패키지 이동
