+++ android/app/src/main/java/com/on/turip/ui/compose/trip/component/TripDetailAppBar.kt
@@ -1,6 +1,7 @@
 package com.on.turip.ui.compose.trip.component
 
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material3.Icon
@@ -8,6 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import com.on.turip.R
 import com.on.turip.ui.compose.designsystem.component.TuripAppBar
 import com.on.turip.ui.compose.designsystem.theme.TuripTheme
@@ -38,7 +40,10 @@ fun TripDetailAppBar(
                 Icon(
                     painter = painterResource(contentFavoriteIconRes),
                     contentDescription = null,
-                    modifier = Modifier.clickable(onClick = onContentFavoriteClick),
+                    modifier =
+                        Modifier
+                            .clickable(onClick = onContentFavoriteClick)
+                            .size(20.dp),
                 )
             }
         },
