+++ android/app/src/main/java/com/on/turip/ui/main/home/HomeFragment.kt
@@ -14,7 +14,7 @@ import com.on.turip.ui.login.LoginActivity
 import com.on.turip.ui.main.home.model.UsersLikeContentModel
 import com.on.turip.ui.search.keywordresult.SearchActivity
 import com.on.turip.ui.search.regionresult.RegionResultActivity
-import com.on.turip.ui.trip.detail.TripDetailActivity
+import com.on.turip.ui.trip.TripDetailActivity
 import dagger.hilt.android.AndroidEntryPoint
 import timber.log.Timber
