+++ android/app/src/main/java/com/on/turip/ui/search/regionresult/RegionResultActivity.kt
@@ -13,7 +13,7 @@ import com.on.turip.ui.common.base.BaseActivity
 import com.on.turip.ui.common.collectOnStarted
 import com.on.turip.ui.common.error.ErrorUiState
 import com.on.turip.ui.login.LoginActivity
-import com.on.turip.ui.trip.detail.TripDetailActivity
+import com.on.turip.ui.trip.TripDetailActivity
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
