--- android/app/src/main/res/layout/bottom_sheet_fragment_favorite_help_information.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <View
-        android:id="@+id/v_bottom_sheet_favorite_top_space"
-        android:layout_width="match_parent"
-        android:layout_height="30dp"
-        android:background="@null"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_bottom_sheet_favorite_help_information"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:orientation="vertical"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/v_bottom_sheet_favorite_top_space"
-        tools:listitem="@layout/item_help_information" />
-
-    <View
-        android:id="@+id/v_bottom_sheet_favorite_bottom_space"
-        android:layout_width="match_parent"
-        android:layout_height="20dp"
-        android:background="@null"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/rv_bottom_sheet_favorite_help_information" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
