--- android/app/src/main/res/layout/item_trip_day.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:paddingTop="16dp"
-    android:paddingBottom="8dp">
-
-    <TextView
-        android:id="@+id/tv_trip_day_day"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginEnd="6dp"
-        android:textAppearance="@style/title3"
-        android:textColor="@color/pure_black_151515"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:text="1일차" />
-
-    <com.google.android.material.divider.MaterialDivider
-        android:id="@+id/div_trip_day_underline"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="4dp"
-        app:dividerColor="@color/gray_200_c1c1c1"
-        app:dividerThickness="2dp"
-        app:layout_constraintEnd_toEndOf="@id/tv_trip_day_day"
-        app:layout_constraintStart_toStartOf="@id/tv_trip_day_day"
-        app:layout_constraintTop_toBottomOf="@id/tv_trip_day_day" />
-</androidx.constraintlayout.widget.ConstraintLayout>
