--- android/app/src/main/res/layout/item_trip_place.xml
@@ -1,164 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/cl_trip_place"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_marginHorizontal="20dp"
-    android:layout_marginBottom="12dp"
-    android:background="@drawable/bg_turip_light_gray_50_stroke_turip_black_06_radius_24dp"
-    android:paddingVertical="20dp">
-
-    <TextView
-        android:id="@+id/tv_trip_place_category"
-        android:layout_width="wrap_content"
-        android:layout_height="0dp"
-        android:layout_marginHorizontal="28dp"
-        android:textAppearance="@style/title3"
-        android:textColor="@color/gray_400_2b2b2b"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_goneMarginTop="20dp"
-        tools:text="관광지" />
-
-    <TextView
-        android:id="@+id/tv_trip_place_name"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="28dp"
-        android:ellipsize="end"
-        android:maxLines="1"
-        android:textAppearance="@style/title1"
-        android:textColor="@color/gray_400_2b2b2b"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="@id/tv_trip_place_category"
-        app:layout_constraintTop_toBottomOf="@id/tv_trip_place_category"
-        tools:text="경포대 해수욕장경포대 해수욕장경포대 해수욕장경포대 해수욕장" />
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/cl_trip_place_time_line"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="12dp"
-        app:layout_constraintEnd_toStartOf="@id/view_trip_place_divider1"
-        app:layout_constraintHorizontal_chainStyle="spread"
-        app:layout_constraintStart_toStartOf="@id/tv_trip_place_name"
-        app:layout_constraintTop_toBottomOf="@id/tv_trip_place_name">
-
-        <ImageView
-            android:id="@+id/iv_trip_place_time_line"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:contentDescription="@null"
-            android:paddingVertical="8dp"
-            android:src="@drawable/ic_youtuber_badge"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:tint="@color/turip_blue_5ac3d5" />
-
-        <TextView
-            android:id="@+id/tv_trip_place_time_line"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:textAppearance="@style/info1"
-            android:textColor="@color/gray_400_2b2b2b"
-            app:layout_constraintBottom_toBottomOf="@id/iv_trip_place_time_line"
-            app:layout_constraintStart_toEndOf="@id/iv_trip_place_time_line"
-            app:layout_constraintTop_toTopOf="@id/iv_trip_place_time_line"
-            tools:text="33:03~" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <View
-        android:id="@+id/view_trip_place_divider1"
-        android:layout_width="1dp"
-        android:layout_height="0dp"
-        android:layout_marginVertical="8dp"
-        android:background="@color/pure_white_ffffff"
-        app:layout_constraintBottom_toBottomOf="@id/cl_trip_place_map"
-        app:layout_constraintEnd_toStartOf="@id/cl_trip_place_map"
-        app:layout_constraintStart_toEndOf="@id/cl_trip_place_time_line"
-        app:layout_constraintTop_toTopOf="@id/cl_trip_place_map" />
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/cl_trip_place_map"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:padding="8dp"
-        app:layout_constraintBottom_toBottomOf="@id/cl_trip_place_time_line"
-        app:layout_constraintEnd_toStartOf="@id/view_trip_place_divider2"
-        app:layout_constraintStart_toEndOf="@id/view_trip_place_divider1"
-        app:layout_constraintTop_toTopOf="@id/cl_trip_place_time_line">
-
-        <ImageView
-            android:id="@+id/iv_trip_place_map"
-            android:layout_width="30dp"
-            android:layout_height="30dp"
-            android:contentDescription="@null"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:src="@drawable/btn_map_link" />
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:text="@string/trip_place_map"
-            android:textAppearance="@style/info1"
-            android:textColor="@color/gray_400_2b2b2b"
-            app:layout_constraintBottom_toBottomOf="@id/iv_trip_place_map"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toEndOf="@id/iv_trip_place_map"
-            app:layout_constraintTop_toTopOf="@id/iv_trip_place_map" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <View
-        android:id="@+id/view_trip_place_divider2"
-        android:layout_width="1dp"
-        android:layout_height="0dp"
-        android:layout_marginVertical="8dp"
-        android:background="@color/pure_white_ffffff"
-        app:layout_constraintBottom_toBottomOf="@id/cl_trip_place_map"
-        app:layout_constraintEnd_toStartOf="@id/cl_trip_place_favorite"
-        app:layout_constraintStart_toEndOf="@id/cl_trip_place_map"
-        app:layout_constraintTop_toTopOf="@id/cl_trip_place_map" />
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/cl_trip_place_favorite"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:paddingVertical="8dp"
-        app:layout_constraintBottom_toBottomOf="@id/view_trip_place_divider2"
-        app:layout_constraintEnd_toEndOf="@id/tv_trip_place_name"
-        app:layout_constraintStart_toEndOf="@id/view_trip_place_divider2"
-        app:layout_constraintTop_toTopOf="@id/view_trip_place_divider2">
-
-        <ImageView
-            android:id="@+id/iv_trip_place_favorite"
-            android:layout_width="18dp"
-            android:layout_height="18dp"
-            android:contentDescription="@null"
-            android:src="@drawable/btn_favorite_red"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:text="@string/trip_place_favorite"
-            android:textAppearance="@style/info1"
-            android:textColor="@color/gray_400_2b2b2b"
-            app:layout_constraintBottom_toBottomOf="@id/iv_trip_place_favorite"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toEndOf="@id/iv_trip_place_favorite"
-            app:layout_constraintTop_toTopOf="@id/iv_trip_place_favorite" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
